feat(sidebar): Add back functionality to mobile sidebar
Introduces a new `onBack` prop to the `SidebarBody` component. This allows consumers to define a callback function that is executed when the "Back" button in the mobile sidebar is pressed. The `App.tsx` file is updated to utilize this new functionality, setting the active tab to "landing" when the sidebar is closed from a sub-view.